+++ runtime/runtime-params-estimator/src/v2/gas_cost.rs
@@ -1,0 +1,77 @@
+use std::cmp::Ordering;
+use std::time::Duration;
+use std::{fmt, ops};
+
+use near_primitives::types::Gas;
+use num_rational::Ratio;
+
+use crate::cases::ratio_to_gas;
+use crate::testbed_runners::GasMetric;
+
+#[derive(Clone, PartialEq, Eq)]
+pub(crate) struct GasCost {
+    /// The smallest thing we are measuring is one wasm instruction, and it
+    /// takes about a nanosecond, so we do need to account for fractional
+    /// nanoseconds here!
+    pub value: Ratio<u64>,
+    pub metric: GasMetric,
+}
+
+impl fmt::Debug for GasCost {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self.metric {
+            GasMetric::ICount => write!(f, "{}i", self.value),
+            GasMetric::Time => fmt::Debug::fmt(&Duration::from_nanos(self.value.to_integer()), f),
+        }
+    }
+}
+
+impl ops::Add for GasCost {
+    type Output = GasCost;
+
+    fn add(self, rhs: GasCost) -> Self::Output {
+        assert_eq!(self.metric, rhs.metric);
+        GasCost { value: self.value + rhs.value, metric: self.metric }
+    }
+}
+
+impl ops::AddAssign for GasCost {
+    fn add_assign(&mut self, rhs: GasCost) {
+        *self = self.clone() + rhs;
+    }
+}
+
+impl ops::Sub for GasCost {
+    type Output = GasCost;
+
+    fn sub(self, rhs: GasCost) -> Self::Output {
+        assert_eq!(self.metric, rhs.metric);
+        GasCost { value: self.value - rhs.value, metric: self.metric }
+    }
+}
+
+impl ops::Div<u64> for GasCost {
+    type Output = GasCost;
+
+    fn div(self, rhs: u64) -> Self::Output {
+        GasCost { value: self.value / rhs, metric: self.metric }
+    }
+}
+
+impl PartialOrd for GasCost {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for GasCost {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.value.cmp(&other.value)
+    }
+}
+
+impl GasCost {
+    pub(crate) fn to_gas(self) -> Gas {
+        ratio_to_gas(self.metric, self.value)
+    }
+}
